commit with medical
There is a now a prototype application of the ising model in modelling the spread of pathogenic diseases. currently working off the assumption that no-one will recover from the disease, only die. Need to play with the probabilities such that I have total control of the program with the metropolis probability refliecting virility of disease and immune strength of the population.